+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfix
  - `select` when `required` now properly check if it is filled
  - style `showUserInput` show `timestamp` instead of `record_id`
+ - properly assign the value for `field_not_found` for custom fields in `data_config`;
 
 # v5.10.1
 ###  New features