+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - add field `internal` to styles: `formUserInput`, `formUserInputLog` and `formUserInputRecord`. If internal is enabled the form data is not shown and can be used only internally.
  - #342 add an option to set a default value for `gender` and `name`(user_name) in style `validation`. If the default value is set the fields will be hidden.
  - #343 add a field `email_activate_email_address` in `Email Templates`. If the field is set, it will be used as an email from where is sent the message and where it can be replied.
+ - #344 add fields in `Email Templates` for sending emails when a profile is deleted. If the fields are filled the email will be sent to the user. If the field `email_delete_profile_email_address_notification_copy` is set, another email will be sent to this address. This functionality is used to notify admin users that user in the system deleted their profile.
 
 ### Bugfix
  - `select` when `required` now properly check if it is filled